Add cast to appease clang-armv7-linux-build-cache (D69471 followup)
http://lab.llvm.org:8011/builders/clang-armv7-linux-build-cache/builds/27075

error: non-constant-expression cannot be narrowed from type 'uint64_t'
(aka 'unsigned long long') to 'size_t' (aka 'unsigned int') in
initializer list [-Wc++11-narrowing]
  return {MappingBuf, getDataSize<FuncRecordTy, Endian>(Record)};

+++ llvm/include/llvm/ProfileData/Coverage/CoverageMapping.h
@@ -743,7 +743,7 @@ Error getFuncNameViaRef(const FuncRecordTy *Record,
 template <class FuncRecordTy, support::endianness Endian>
 StringRef getCoverageMappingOutOfLine(const FuncRecordTy *Record,
                                       const char *MappingBuf) {
-  return {MappingBuf, getDataSize<FuncRecordTy, Endian>(Record)};
+  return {MappingBuf, size_t(getDataSize<FuncRecordTy, Endian>(Record))};
 }
 
 /// Advance to the next out-of-line coverage mapping and its associated
